Modified coins_table_migration: id

+++ database/migrations/2021_06_20_151757_create_deposits_table.php
@@ -17,7 +17,7 @@ class CreateDepositsTable extends Migration
             $table->increments('id');
             $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->bigInteger('coin_id')->unsigned();
+            $table->integer('coin_id')->unsigned();
             $table->foreign('coin_id')->references('id')->on('coins')->onDelete('cascade');
             $table->bigInteger('package_id')->unsigned();
             $table->foreign('package_id')->references('id')->on('packages')->onDelete('cascade');
